Move preprocess args to schema
