+++ src/guidellm/schemas/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Schemas for CLI commands."""
+
+from .preprocess import PreprocessDatasetArgs, default_kind
+
+__all__ = ["PreprocessDatasetArgs", "default_kind"]
